requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-default-group==1.2.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
crypto==1.4.1
des==1.0.5
dj-config-url==0.1.1
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==3.1.1
kombu==5.0.2
Naked==0.1.31
postgres==3.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycrypto==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
shellescape==3.8.1
six==1.15.0
sqlparse==0.4.1
tabulate==0.8.7
task==0.2.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
